requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
pikepdf>=8.0.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
